+++ src/pages/search/Search.jsx
@@ -1,0 +1,5 @@
+const Search = () => {
+    return <></>
+}
+
+export default Search
